pyproject.toml
[project]
name = "crew"
version = "0.1.0"
description = "ai-trading-crew using crewAI"
authors = [{ name = "Antigravity", email = "antigravity@gemini.google.com" }]
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]",
    "onnxruntime",
    "numpy",
    "yfinance",
    "pandas-datareader",
    "pandas-market-calendars",
    "nixtla",
    "matplotlib",
    "ta-lib",
    "linkup-sdk",
    "crawl4ai",
    "openpyxl",
    "requests",
    "seaborn",
    "yahooquery",
]

[project.scripts]
crew = "crew.main:run"
run_crew = "crew.main:run"
train = "crew.main:train"
replay = "crew.main:replay"
test = "crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
